jenconn-api: change api structure
