+++ cmd/demo/main.go
@@ -2,8 +2,14 @@ package main
 
 import (
 	"marketplace/api"
+	db_component "marketplace/internal/db-component"
 )
 
 func main() {
-	api.RunApi()
+	ctx, rdb := db_component.ConnectToRedis()
+	defer rdb.Close()
+
+	api := api.NewAPI(ctx, rdb)
+
+	api.Run()
 }
